+++ src/app/integrations/utils.ts
@@ -29,6 +29,7 @@ export enum INTEGRATIONS_KEYS {
     COHERE = 'cohere',
     DEEPSEEK = 'deepseek',
     OLLAMA = 'ollama',
+    MCP = 'mcp',
     VLLM = 'vllm'
 }
 
@@ -46,6 +47,7 @@ export enum APPS_TITLE {
 
 export enum INTEGRATIONS_CATEGORIES {
     AI = 'ai',
+    MCP = 'mcp',
     AUTOMATION = "automation",
     CHANNEL = "channel",
     COMMUNICATION = 'communication',
@@ -60,6 +62,7 @@ export enum INTEGRATIONS_CATEGORIES {
 
 export const CATEGORIES_LIST = [
     { type: INTEGRATIONS_CATEGORIES.AI, name: "Artificial Intelligence" },
+    { type: INTEGRATIONS_CATEGORIES.MCP, name: "MCP" },
     { type: INTEGRATIONS_CATEGORIES.CHANNEL, name: "Channels" },
     //{ type: INTEGRATIONS_CATEGORIES.AUTOMATION, name: "Automation" },
     //{ type: INTEGRATIONS_CATEGORIES.COMMUNICATION, name: "Communication" },
@@ -114,6 +117,7 @@ export const INTEGRATION_LIST_ARRAY = [
     { name: "Groq",                  category: INTEGRATIONS_CATEGORIES.AI,               key: INTEGRATIONS_KEYS.GROQ,            src_icon: "assets/img/int/groq-icon.svg",                   src_logo: "assets/img/int/groq_logo.svg",               pro: true,  plan: 'Pro' },
     { name: "Cohere",                category: INTEGRATIONS_CATEGORIES.AI,               key: INTEGRATIONS_KEYS.COHERE,          src_icon: "assets/img/int/cohere_icon.svg",                 src_logo: "assets/img/int/cohere_logo.svg",             pro: true,  plan: 'Pro' },
     { name: "Ollama",                category: INTEGRATIONS_CATEGORIES.AI,               key: INTEGRATIONS_KEYS.OLLAMA,          src_icon: "assets/img/int/ollama-icon.svg",                 src_logo: "assets/img/int/ollama-logo.png",             pro: true,  plan: 'Pro' },
+    { name: "MCP Servers",           category: INTEGRATIONS_CATEGORIES.MCP,              key: INTEGRATIONS_KEYS.MCP,             src_icon: "assets/img/int/mcp-icon.png",                    src_logo: "assets/img/int/mcp-logo.png",                pro: true,  plan: 'Pro' },
     { name: "vLLM",                  category: INTEGRATIONS_CATEGORIES.AI,               key: INTEGRATIONS_KEYS.VLLM,            src_icon: "assets/img/int/vllm-icon.png",                   src_logo: "assets/img/int/vllm-logo.png",               pro: true,  plan: 'Pro' },
     { name: "Deepseek",              category: INTEGRATIONS_CATEGORIES.AI,               key: INTEGRATIONS_KEYS.DEEPSEEK,        src_icon: "assets/img/int/deepseek-icon.svg",               src_logo: "assets/img/int/deepseek-logo.svg",           pro: true,  plan: 'Pro' },
     { name: "Qapla'",                category: INTEGRATIONS_CATEGORIES.ECOMMERCE,        key: INTEGRATIONS_KEYS.QAPLA,           src_icon: "assets/img/int/qapla-icon.jpg",                  src_logo: "assets/img/int/qapla-logo.png",              pro: true,  plan: 'Pro' },
@@ -215,6 +219,16 @@ export class OllamaIntegration extends Integration {
             url: null,
             token: null,
             models: []
+        }
+    }
+}
+
+export class McpIntegration extends Integration {
+    constructor() {
+        super();
+        this.name = INTEGRATIONS_KEYS.MCP;
+        this.value = {
+            servers: []
         }
     }
 }